# planner: optimize return type of partial sum for tiflash