--- a/pkg/expression/aggregation/base_func.go
+++ b/pkg/expression/aggregation/base_func.go
@@ -244,34 +244,6 @@ func (a *baseFuncDesc) typeInfer4Sum(ctx expression.EvalContext) {
 	types.SetBinChsClnFlag(a.RetTp)
 }
 
-// TypeInfer4AvgSum infers the type of sum from avg, which should extend the precision of decimal
-// compatible with mysql.
-func (a *baseFuncDesc) TypeInfer4AvgSum(isTiFlash bool, avgRetTp *types.FieldType, avgArgRetTp *types.FieldType) {
-	if avgRetTp.GetType() == mysql.TypeNewDecimal {
-		if isTiFlash && avgArgRetTp.GetType() == mysql.TypeNewDecimal {
-			// For tiflash, different precision use different type to compute decimal.
-			// For precision in (18, 38], will use Int128.
-			// For precision in (38, 65], will use Int256. And Int256 is much slower than Int128.
-			// So, setting flen more accurately can lead to significant performance improvements.
-			if avgArgRetTp.GetFlen() <= 9 {
-				a.RetTp.SetFlen(18)
-			} else if avgArgRetTp.GetFlen() <= 18 {
-				a.RetTp.SetFlen(min(avgArgRetTp.GetFlen()+22, 38))
-			} else {
-				a.RetTp.SetFlen(min(mysql.MaxDecimalWidth, avgArgRetTp.GetFlen()+22))
-			}
-			if avgArgRetTp.GetDecimal() > 0 {
-				a.RetTp.SetDecimal(avgArgRetTp.GetDecimal())
-			} else {
-				// Normally cannot reach here for decimal type.
-				a.RetTp.SetDecimal(avgRetTp.GetDecimal())
-			}
-		} else {
-			a.RetTp.SetFlen(min(mysql.MaxDecimalWidth, a.RetTp.GetFlen()+22))
-		}
-	}
-}
-
 // TypeInfer4FinalCount infers the type of sum agg which is rewritten from final count agg run on MPP mode.
 func (a *baseFuncDesc) TypeInfer4FinalCount(finalCountRetType *types.FieldType) {
 	a.RetTp = finalCountRetType.Clone()