--- a/pkg/expression/aggregation/base_func.go
+++ b/pkg/expression/aggregation/base_func.go
@@ -250,6 +250,10 @@ func (a *baseFuncDesc) TypeInfer4AvgSum(ctx expression.EvalContext, avgRetType *
 	if a.Name != ast.AggFuncSum {
 		return errors.Errorf("expect sum func, but got %s", a.Name)
 	}
+	// Handling column and scalar function differently to avoid breaking a MySQL compatible issue.
+	// Check: https://github.com/pingcap/tidb/blob/67edd7d8f73de399bd72490d449d1dede1ee637b/pkg/executor/test/tiflashtest/tiflash_test.go#L887
+	// For avg(div(col1, col2)), the scale of div result should be same as the scale of avg, which has been increased by 4, to make sure the result is compatible with MySQL.
+	// But for avg(col1), there is no need to increase the result scale of partial sum, because there is no complex scale upgrade for a simple column.
 	if _, ok := a.Args[0].(*expression.Column); ok {
 		a.typeInfer4Sum(ctx)
 	} else {