--- a/pkg/planner/core/task.go
+++ b/pkg/planner/core/task.go
@@ -1611,7 +1611,9 @@ func BuildFinalModeAggregation(
 				// we must call deep clone in this case, to avoid sharing the arguments.
 				sumAgg := aggFunc.Clone()
 				sumAgg.Name = ast.AggFuncSum
-				sumAgg.TypeInfer4AvgSum(isMPPTask, aggFunc.RetTp, aggFunc.Args[0].GetType(ectx))
+				if err := sumAgg.TypeInfer(sctx.GetExprCtx()); err != nil {
+					// todo err
+				}
 				partial.Schema.Columns[partialCursor-1].RetType = sumAgg.RetTp
 				partial.AggFuncs = append(partial.AggFuncs, cntAgg, sumAgg)
 			} else if aggFunc.Name == ast.AggFuncApproxCountDistinct || aggFunc.Name == ast.AggFuncGroupConcat {
@@ -1686,7 +1688,9 @@ func (p *basePhysicalAgg) convertAvgForMPP() *PhysicalProjection {
 			// insert a sum(column)
 			avgSum := aggFunc.Clone()
 			avgSum.Name = ast.AggFuncSum
-			avgSum.TypeInfer4AvgSum(true, aggFunc.RetTp, aggFunc.Args[0].GetType(exprCtx.GetEvalCtx()))
+			if err := avgSum.TypeInfer(exprCtx); err != nil {
+				// todo err
+			}
 			newAggFuncs = append(newAggFuncs, avgSum)
 			avgSumCol := &expression.Column{
 				UniqueID: p.schema.Columns[i].UniqueID,