<!--

Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?
<!--

Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and
linking the relevant issues via the "close" or "ref".

For more info, check https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/contribute-code.html#referring-to-an-issue.

-->

Issue Number: close #59708

Problem Summary: 
the type infer of `avg(col_decimal_15_2)`:
  1. In `plan builder`, call [typeInfer4Avg](https://github.com/pingcap/tidb/blob/62165a41faa800def93a0582dad918ad669ca435/pkg/expression/aggregation/base_func.go#L138), the return type of avg becomes `decimal(19, 6)`
  2. In `task.go`, trying to convert normal `avg` to mpp `avg`, which will be spilit to `count() and sum(col_decimal_15_2)`(1 phase) and `sum/count`(2 phase). And the return type of 1 phase `sum(col_decimal_15_2)` is `decimal(41, 6)` because of calling [TypeInfer4AvgSum](https://github.com/pingcap/tidb/blob/62165a41faa800def93a0582dad918ad669ca435/pkg/expression/aggregation/base_func.go#L249)

`decimal(41, 6)` will use Int256 to compute, which is slow. So we need to optimize type infer.

### What changed and how does it work?

1. optimize the partial sum of avg for tiflash. So appropriate type will be used to compute sum.
2. the final avg ret type should not be changed.

Workload: tpch50G
Query: `explain analyze select avg(l_quantity) from lineitem group by l_orderkey;`
performance: 4-> 2.9 (**+26%**)
before:
```
mysql> mysql> explain analyze select avg(l_quantity) avgres from lineitem group by l_orderkey having avgres > 100000;
+--------------------------------------+--------------+-----------+--------------+----------------+----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+-------------------------------------------------------------------------------------------------------------------------------------+---------+------+
| id                                   | estRows      | actRows   | task         | access object  | execution info                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       | operator info                                                                                                                       | memory  | disk |
+--------------------------------------+--------------+-----------+--------------+----------------+----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+-------------------------------------------------------------------------------------------------------------------------------------+---------+------+
| TableReader_65                       | 59126579.20  | 0         | root         |                | time:3.98s, open:1.53ms, close:6.64µs, loops:1, RU:113028.94, cop_task: {num: 2, max: 0s, min: 0s, avg: 0s, p95: 0s, copr_cache_hit_ratio: 0.00}                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     | MppVersion: 3, data:ExchangeSender_64                                                                                               | 1.50 KB | N/A  |
| └─ExchangeSender_64                  | 59126579.20  | 0         | mpp[tiflash] |                | tiflash_task:{proc max:3.98s, min:3.92s, avg: 3.95s, p80:3.98s, p95:3.98s, iters:0, tasks:2, threads:32}                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             | ExchangeType: PassThrough                                                                                                           | N/A     | N/A  |
|   └─Selection_63                     | 59126579.20  | 0         | mpp[tiflash] |                | tiflash_task:{proc max:3.98s, min:3.92s, avg: 3.95s, p80:3.98s, p95:3.98s, iters:0, tasks:2, threads:32}                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             | gt(Column#17, 100000)                                                                                                               | N/A     | N/A  |
|     └─Projection_57                  | 73908224.00  | 75000000  | mpp[tiflash] |                | tiflash_task:{proc max:3.94s, min:3.88s, avg: 3.91s, p80:3.94s, p95:3.94s, iters:4096, tasks:2, threads:32}                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          | div(Column#17, cast(case(eq(Column#31, 0), 1, Column#31), decimal(20,0) BINARY))->Column#17                                         | N/A     | N/A  |
|       └─HashAgg_58                   | 73908224.00  | 75000000  | mpp[tiflash] |                | tiflash_task:{proc max:2.82s, min:2.81s, avg: 2.82s, p80:2.82s, p95:2.82s, iters:4096, tasks:2, threads:32}                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          | group by:test.lineitem.l_orderkey, funcs:sum(Column#32)->Column#31, funcs:sum(Column#33)->Column#17, stream_count: 16               | N/A     | N/A  |
|         └─ExchangeReceiver_60        | 73908224.00  | 75000093  | mpp[tiflash] |                | tiflash_task:{proc max:2.19s, min:2.15s, avg: 2.17s, p80:2.19s, p95:2.19s, iters:7226, tasks:2, threads:32}, tiflash_wait: {pipeline_queue_wait: 511ms}, tiflash_network: {inner_zone_receive_bytes: 561417293}                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      | stream_count: 16                                                                                                                    | N/A     | N/A  |
|           └─ExchangeSender_59        | 73908224.00  | 75000093  | mpp[tiflash] |                | tiflash_task:{proc max:2.61s, min:0s, avg: 1.3s, p80:2.61s, p95:2.61s, iters:1536, tasks:2, threads:144}, tiflash_network: {inner_zone_send_bytes: 561417293}                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        | ExchangeType: HashPartition, Compression: FAST, Hash Cols: [name: test.lineitem.l_orderkey, collate: binary], stream_count: 16      | N/A     | N/A  |
|             └─HashAgg_55             | 73908224.00  | 75000093  | mpp[tiflash] |                | tiflash_task:{proc max:1.9s, min:0s, avg: 951.2ms, p80:1.9s, p95:1.9s, iters:1536, tasks:2, threads:144}                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             | group by:test.lineitem.l_orderkey, funcs:count(test.lineitem.l_quantity)->Column#32, funcs:sum(test.lineitem.l_quantity)->Column#33 | N/A     | N/A  |
|               └─TableFullScan_35     | 300005811.00 | 300005811 | mpp[tiflash] | table:lineitem | tiflash_task:{proc max:786.5ms, min:0s, avg: 393.2ms, p80:786.5ms, p95:786.5ms, iters:4757, tasks:2, threads:144}, tiflash_wait: {pipeline_queue_wait: 772ms}, tiflash_scan:{mvcc_input_rows:0, mvcc_input_bytes:0, mvcc_output_rows:0, local_regions:217, remote_regions:0, tot_learner_read:6ms, region_balance:{instance_num: 2, max/min: 109/108=1.009259}, delta_rows:0, delta_bytes:0, segments:434, stale_read_regions:0, tot_build_snapshot:0ms, tot_build_bitmap:25ms, tot_build_inputstream:226ms, min_local_stream:323ms, max_local_stream:780ms, dtfile:{data_scanned_rows:300005811, data_skipped_rows:682517, mvcc_scanned_rows:0, mvcc_skipped_rows:0, lm_filter_scanned_rows:0, lm_filter_skipped_rows:0, tot_rs_index_check:27ms, tot_read:5806ms}} | keep order:false                                                                                                                    | N/A     | N/A  |
+--------------------------------------+--------------+-----------+--------------+----------------+----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+-------------------------------------------------------------------------------------------------------------------------------------+---------+------+
9 rows in set (4.00 sec)

```

after:
```
mysql> mysql> explain analyze select avg(l_quantity) avgres from lineitem group by l_orderkey having avgres > 100000;
+--------------------------------------+--------------+-----------+--------------+----------------+--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+-------------------------------------------------------------------------------------------------------------------------------------+---------+------+
| id                                   | estRows      | actRows   | task         | access object  | execution info                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               | operator info                                                                                                                       | memory  | disk |
+--------------------------------------+--------------+-----------+--------------+----------------+--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+-------------------------------------------------------------------------------------------------------------------------------------+---------+------+
| TableReader_65                       | 59126579.20  | 0         | root         |                | time:2.91s, open:1.53ms, close:5.11µs, loops:1, RU:102874.94, cop_task: {num: 2, max: 0s, min: 0s, avg: 0s, p95: 0s, copr_cache_hit_ratio: 0.00}                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             | MppVersion: 3, data:ExchangeSender_64                                                                                               | 1.50 KB | N/A  |
| └─ExchangeSender_64                  | 59126579.20  | 0         | mpp[tiflash] |                | tiflash_task:{proc max:2.91s, min:2.88s, avg: 2.89s, p80:2.91s, p95:2.91s, iters:0, tasks:2, threads:32}                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     | ExchangeType: PassThrough                                                                                                           | N/A     | N/A  |
…